backend/requirements.txt
flask
pymongo
python-dotenv
scikit-learn
xgboost
flask-mail
flask-cors
